Finishing up the mailer.js
Changed certificationmailer to general mailer so it can be used for mailing event promotion too.
Began process of editing event details to add the certification option if organizer
Certifier may be replaced with a pop-up overlay later on.
Still need to:
Implement database integration - pull list of user emails for event promotion, grab associated folks for each event to mail certificates to.